Refactor: Remove ThemeToggle component and related styles
Refactor: Update UserWorkspaces component to streamline navigation

Enhancement: Modify WelcomeHeader to adjust user role display

Enhancement: Add loading state to AuthContext for better user experience

Fix: Clean up index.css by removing unnecessary comments

Feature: Revamp GroupPage with tab navigation for better organization

Cleanup: Remove unused mock user data from UserDashboard

Enhancement: Update WorkspacePage to include member count in group cards

Fix: Clean up SignInPage by removing commented-out code

Refactor: Simplify darkTheme and lightTheme by removing accent colors

Refactor: Remove deprecated theme.ts file and update imports

Feature: Add BackButton component for improved navigation

Feature: Create Chats button with badge for notifications

Feature: Implement NotificationsButton with badge functionality

Refactor: Move ThemeToggle to Buttons directory for better organization
