+++ src/App.css
@@ -40,3 +40,7 @@
 .read-the-docs {
   color: #888;
 } */
+
+&::-webkit-scrollbar {
+  display: none;
+}
