+++ src/components/Buttons/BackButton.tsx
@@ -1,57 +1,67 @@
 import React from "react";
 import { IconButton, Tooltip } from "@mui/material";
-import { DarkMode, LightMode } from "@mui/icons-material";
-import { useTheme } from "../theme";
+import { ArrowBack } from "@mui/icons-material";
+import { useTheme } from "../../theme";
 
-interface ThemeToggleProps {
+interface BackButtonProps {
   size?: "small" | "medium" | "large";
   showTooltip?: boolean;
   variant?: "default" | "contained" | "outlined";
 }
 
-export const ThemeToggle: React.FC<ThemeToggleProps> = ({
+export const BackButton: React.FC<BackButtonProps> = ({
   size = "medium",
   showTooltip = true,
   variant = "default",
 }) => {
-  const { mode, toggleTheme } = useTheme();
+  const theme = useTheme();
 
   const getBackgroundSx = () => {
     if (variant === "contained") {
       return {
-        backgroundColor: mode === "light" ? "primary.main" : "secondary.main",
+        backgroundColor:
+          theme.mode === "light" ? "primary.main" : "secondary.main",
         color:
-          mode === "light" ? "primary.contrastText" : "secondary.contrastText",
+          theme.mode === "light"
+            ? "primary.contrastText"
+            : "secondary.contrastText",
         "&:hover": {
-          backgroundColor: mode === "light" ? "primary.dark" : "secondary.dark",
+          backgroundColor:
+            theme.mode === "light" ? "primary.dark" : "secondary.dark",
         },
       };
     }
     if (variant === "outlined") {
       return {
         border: "1px solid",
-        borderColor: mode === "light" ? "primary.main" : "secondary.main",
-        color: mode === "light" ? "primary.main" : "secondary.main",
+        borderColor: theme.mode === "light" ? "primary.main" : "secondary.main",
+        color: theme.mode === "light" ? "primary.main" : "secondary.main",
         "&:hover": {
           backgroundColor:
-            mode === "light" ? "primary.light" : "secondary.light",
+            theme.mode === "light" ? "primary.light" : "secondary.light",
           opacity: 0.1,
         },
       };
     }
     return {
-      color: mode === "light" ? "text.primary" : "text.primary",
+      color: theme.mode === "light" ? "text.primary" : "text.primary",
       "&:hover": {
         backgroundColor: "action.hover",
       },
     };
   };
 
-  const toggleButton = (
+  const handleBackClick = () => {
+    setTimeout(() => {
+      window.history.back();
+    }, 1000);
+  };
+
+  const backButton = (
     <IconButton
-      onClick={toggleTheme}
+      onClick={handleBackClick}
       size={size}
-      aria-label={`Switch to ${mode === "light" ? "dark" : "light"} mode`}
+      aria-label="Go-back"
       sx={{
         transition: "all 0.3s ease-in-out",
         ...getBackgroundSx(),
@@ -69,17 +79,12 @@ export const ThemeToggle: React.FC<ThemeToggleProps> = ({
         },
       }}
     >
-      {mode === "light" ? <DarkMode /> : <LightMode />}
+      <ArrowBack />
     </IconButton>
   );
 
   if (showTooltip) {
-    return (
-      <Tooltip title={`Switch to ${mode === "light" ? "dark" : "light"} mode`}>
-        {toggleButton}
-      </Tooltip>
-    );
+    return <Tooltip title="Go back">{backButton}</Tooltip>;
   }
-
-  return toggleButton;
+  return backButton;
 };
