+++ src/contexts/Authcontext.tsx
@@ -7,6 +7,7 @@ type AuthContextType = {
   user_id: string | null;
   name: string | null;
   email: string | null;
+  loading: boolean;
   setAuth: (user_id: string, role: string) => void;
   clearAuth: () => void;
 };
@@ -18,44 +19,42 @@ export const AuthProvider = ({ children }: { children: ReactNode }) => {
   const [user_id, setUser_id] = useState<string | null>(null);
   const [name, setName] = useState<string | null>(null);
   const [email, setEmail] = useState<string | null>(null);
-
+  const [loading, setLoading] = useState<boolean>(true); 
 
   const fetchUserData = async (userId: string) => {
     try {
+      setLoading(true); 
       const response = await axiosInstance.get(`/auth/get-user/${userId}`);
       console.log("Fetched user data:", response.data);
 
       if (response.data.success) {
         const user = response.data.user;
-        // setRole(user.role);
         setName(user.name);
         setEmail(user.email);
       } else {
-        console.error('Failed to fetch user data');
+        console.error("Failed to fetch user data");
       }
     } catch (error) {
-      console.error('Error fetching user data:', error);
+      console.error("Error fetching user data:", error);
+    } finally {
+      setLoading(false); 
     }
   };
 
-
-  useEffect(()=>{
+  useEffect(() => {
     const storedUserId = localStorage.getItem("user_id");
     const storedRole = localStorage.getItem("role");
     if (!storedUserId) {
       clearAuth();
+      setLoading(false); 
       return;
     }
     setUser_id(storedUserId);
     setRole(storedRole);
     fetchUserData(storedUserId);
-    // console.log("Auth context initialized with role:", storedRole, "and user_id:", storedUserId);
   }, []);
 
-
-
-
-  const setAuth = (newUserId: string , newRole: string) => {
+  const setAuth = (newUserId: string, newRole: string) => {
     localStorage.setItem("user_id", newUserId);
     localStorage.setItem("role", newRole);
     setUser_id(newUserId);
@@ -73,7 +72,9 @@ export const AuthProvider = ({ children }: { children: ReactNode }) => {
   };
 
   return (
-    <AuthContext.Provider value={{ role, user_id, name, email, setAuth, clearAuth }}>
+    <AuthContext.Provider
+      value={{ role, user_id, name, email, loading, setAuth, clearAuth }}
+    >
       {children}
     </AuthContext.Provider>
   );
