+++ src/theme/index.ts
@@ -6,5 +6,4 @@ export type { ThemeMode } from './themes';
 
 // Context and components
 export { ThemeProvider, useTheme } from '../contexts/ThemeContext';
-export { ThemeToggle } from '../components/ThemeToggle';
-export { ThemeDemo } from '../components/ThemeDemo';
+export { ThemeToggle } from '../components/Buttons/ThemeToggle';
